Added custom Exceptions

+++ HR.LeaveManagement.Application/Featues/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -22,7 +23,10 @@ namespace HR.LeaveManagement.Application.Featues.LeaveType.Commands.DeleteLeaveT
 		{
 			var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
-			// Verify record exists
+			if(leaveTypeToDelete is null)
+			{
+				throw new NotFoundException(nameof(LeaveType), request.Id);
+			}
 
 			await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
